# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Profile server",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/orders/cinema/{id}": {
            "get": {
                "description": "Add Choose Cinema tikers",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order Tiket"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Select Movie Tiket",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search Date Cinema",
                        "name": "searchDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search Time Cinema",
                        "name": "searchTime",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search Location Cinema",
                        "name": "searchLocation",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "$ref": "#/definitions/models.MoviesCinema"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/profile": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "$ref": "#/definitions/models.Profile"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete profiles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "summary": "Profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "$ref": "#/definitions/models.Profile"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update Movies",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profile"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Update First Name",
                        "name": "first_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Update Last Name",
                        "name": "last_name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Update Phone_Number",
                        "name": "phone_number",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Update Image",
                        "name": "image",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Update Email",
                        "name": "email",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Update Password",
                        "name": "password",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/controllers.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "results": {
                                            "$ref": "#/definitions/models.Profile"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "page_info": {},
                "results": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "models.MoviesCinema": {
            "type": "object",
            "properties": {
                "cinema": {
                    "type": "string"
                },
                "cinema_date": {
                    "type": "string"
                },
                "cinema_location": {
                    "type": "string"
                },
                "cinema_time": {
                    "type": "string"
                },
                "genre": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "tittle": {
                    "type": "string"
                }
            }
        },
...
        "models.Profile": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "salah@mail.com"
                },
                "first_name": {
                    "type": "string",
                    "example": "Salah"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string",
                    "example": "salah.jpg"
                },
                "last_name": {
                    "type": "string",
                    "example": "Alaudin"
                },
                "password": {
                    "type": "string",
                    "example": "Salah1!"
                },
                "phone_number": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}